[WIP] CFG property page for Peview  (#101)
* [peview] Add cfg function list for CFG-aware PE.

* [peview] Add symbol name resolving for Cfg prop page

* [peview] Clean up Cfg entry parsing with proper data structures

* [peview] List CFG whitelist by names

* [peview] Change cfg lockdown bit to SuppressedCall

That how MS call it. (see RtlpGuardAllowSuppressedCall or RtlGuardCheckLongJumpTarget in ntdll)

* [peview] Add entry id column for CFG property page

* [peview] Fix bug when loading symbols for cfg list

The symbol provider was loaded with (ImageBase, SizeOfCode) infos, not
(ImageBase, SizeOfImage).

* [peview] Clarify code in CfgProp and fix typo in cfg item data struct

* [peview] Decref symbol strings after setting the function name

* [peview] Add PhGetMappedImageCfg and PhGetMappedImageCfgEntry in mapimg

Those helpers can parse and return CFG configuration and entries in a
mapped pe.

* [peview] Tidy symbol provider init in a separate function

The code for symbol provider loading is party ripped from mainwnd.c. It
could be a good idea to provide a PhLoadDbgHelpFromCommonLocations (like
Windows Kits) in symprv.

* [peview] Add cfg configuration in load config property sheet