Fix process private/shared commit statistics
Currently it shows the information only for processes in a job.
And actually it is not correct information at all.
Because it contains grouped commit statistics for all processes in the job.

We will use `VM_COUNTERS_EX2` to obtain correct information for process on modern systems.
And use job info as fall-back for older systems.