Update comments
Note: Assigning the constant value -1 to variables or array elements  of type ULONG is not allowed in constant expressions or initializer lists without an explicit cast. C++11 and later generate multiple warnings that this assignment is not safe or portable. We instead use a LONG[] for the table since it allows -1 directly and safely, holds all positive values  that we need, and resolves the narrowing conversion errors.